Create StreamUserControl compatible with rtsp and rtp
