+++ samples/GStreamerWPF/GStreamerD3DSampleCore/GStreamerD3DSampleCore.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>WinExe</OutputType>
+    <OutputType>Exe</OutputType>
     <TargetFramework>net6.0-windows</TargetFramework>
     <Nullable>enable</Nullable>
     <UseWPF>true</UseWPF>
@@ -9,12 +9,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="GstSharp" Version="1.20.2.1" />
     <PackageReference Include="log4net" Version="2.0.14" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\D3D11Scene\D3D11Scene.vcxproj" />
-    <ProjectReference Include="..\..\..\src\Interop.D3DImageEx\Interop.D3DImageEx.vcxproj" />
+    <ProjectReference Include="..\WPFStream\WPFStream.csproj" />
   </ItemGroup>
 </Project>
